account creation system via github and token creation with protected routes

+++ src/routes/memories.ts
@@ -1,10 +1,19 @@
 import { FastifyInstance } from "fastify";
 import { prisma } from "../lib/prisma";
 import { z } from 'zod'
+import { request } from "http";
 
 export async function memoriesRoutes(app:FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub
+      },
       orderBy: {
         createdAt: 'asc'
       }
@@ -19,18 +28,22 @@ export async function memoriesRoutes(app:FastifyInstance) {
     })
   })
 
-  app.get('/memories:id', async (request) => {
+  app.get('/memories:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
 
     const { id } = paramsSchema.parse(request.params)
 
-    const memory = prisma.memory.findUniqueOrThrow({
+    const memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id
       }
     })
+
+    if(!memory.isPublic && memory.userId !== request.user.sub){
+      return reply.status(401)
+    }
 
     return memory
   })
@@ -51,14 +64,14 @@ export async function memoriesRoutes(app:FastifyInstance) {
         conteudo,
         coverUrl,
         isPublic,
-        userId: '27e9dc75-6534-468d-9309-67f8fc590f0f'
+        userId: request.user.sub
       }
     })
 
     return memory
   })
 
-  app.put('/memories:id', async (request) => {
+  app.put('/memories:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
@@ -73,7 +86,17 @@ export async function memoriesRoutes(app:FastifyInstance) {
 
     const { conteudo, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id
+      }
+    })
+
+    if(memory.userId !== request.user.sub){
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id
       }, 
@@ -87,12 +110,22 @@ export async function memoriesRoutes(app:FastifyInstance) {
     return memory
   })
   
-  app.delete('/memories:id', async (request) => {
+  app.delete('/memories:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id
+      }
+    })
+
+    if(memory.userId !== request.user.sub){
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
